apply resolver to tableRow, ensure correct fields to validate
